[wip] Refactor into utility factories
